requirements.txt
praisonaiagents>=0.0.73
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.0
pillow>=10.2.0
requests>=2.31.0
python-multipart>=0.0.9
motor>=3.3.2
aiohttp>=3.9.3
psutil>=5.9.8
markdown>=3.5.2
pygments>=2.17.2
rich>=13.7.0 